# server/docs/openapi/spec.yaml
...
openapi: 3.1.1
info:
  title: Forms Portal API
  version: 0.1.0
  summary: API Documentation for the Forms Portal
  contact:
    name: OSDG
servers:
  - url: /api
    description: Production
...
security:
  - cookieAuth: []
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List groups
      description: Retrieves all groups owned by the current user.
      operationId: listGroups
      parameters:
        - name: sort
          in: query
          description: Sort criteria.
          schema:
            type: string
            enum: [created, name, type]
            default: created
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: A paginated list of groups.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Groups]
      summary: Create group
      description: Creates a new user group with the current user as the owner.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/groupId'
    get:
      tags: [Groups]
      summary: Get group details
      description: Retrieves detailed information about a specific group. Requires ownership.
      operationId: getGroup
      responses:
        '200':
          description: Group details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      tags: [Groups]
      summary: Update group
      description: Updates a group's name or description. Requires ownership.
      operationId: updateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdate'
      responses:
        '200':
          description: Group updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Groups]
      summary: Delete group
      description: Permanently deletes a group. Requires ownership.
      operationId: deleteGroup
      responses:
        '204':
          description: Group deleted successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /groups/{groupId}/domain:
    parameters:
      - $ref: '#/components/parameters/groupId'
    put:
      tags: [Groups]
      summary: Update group domain
      description: Updates the domain for a group of 'DOMAIN' type. Requires ownership.
      operationId: updateGroupDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDomainUpdate'
      responses:
        '200':
          description: Group domain updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/groupId'
    post:
      tags: [Groups]
      summary: Add group member
      description: Adds a user to a 'LIST' type group. Requires group ownership.
      operationId: addGroupMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '204':
          description: Member added successfully.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /groups/{groupId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/groupId'
      - $ref: '#/components/parameters/userId'
    delete:
      tags: [Groups]
      summary: Remove group member
      description: Removes a user from a 'LIST' type group. Requires group ownership.
      operationId: removeGroupMember
      responses:
        '204':
          description: Member removed successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: cookie
      name: session
  parameters:
...
    groupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: ulid
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: ulid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    GroupType:
      type: string
      enum: [domain, list]
...
    Group:
      type: object
      required:
        - id
        - owner
        - name
        - type
      properties:
        id:
          type: string
          format: ulid
        owner:
          type: string
          format: ulid
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/GroupType'
        domain:
          type: string
          format: fqdn
          description: Present if type is 'domain'.
        members:
          type: array
          items:
            type: string
            format: email
          description: Array of user emails. Present if type is 'list'.
    GroupCreate:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/GroupType'
        domain:
          type: string
          format: fqdn
          description: Required if type is 'domain'.
        members:
          type: array
          items:
            type: string
            format: email
          description: Array of user emails. Only applicable if type is 'list'.
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    GroupDomainUpdate:
      type: object
      required:
        - domain
      properties:
        domain:
          type: string
          format: fqdn
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      required:
        - offset
        - limit
        - total
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer